Restructure cortex services to delineate tooling type
